Reconstruct second color interpolation channel

+++ src/Yajima/y_runtime.c
@@ -1540,7 +1540,7 @@ void FUN_0045c530(int param_1);
 void FUN_0045c8c0(int param_1,int param_2);
 void FUN_0045cdd0(int param_1,long param_2);
 void FUN_0045ce90(int param_1);
-void FUN_0045d990(int param_1,long param_2);
+void FUN_0045d990(int param_1,int param_2);
 void FUN_0045e7b0(int param_1);
 void FUN_0045e7e0(void);
 void FUN_0045ea80(int param_1);
@@ -2956,7 +2956,7 @@ void FUN_0045c530(int param_1);
 void FUN_0045c8c0(int param_1,int param_2);
 void FUN_0045cdd0(int param_1,long param_2);
 void FUN_0045ce90(int param_1);
-void FUN_0045d990(int param_1,long param_2);
+void FUN_0045d990(int param_1,int param_2);
 void FUN_0045e7b0(int param_1);
 void FUN_0045e7e0(void);
 void FUN_0045ea80(int param_1);
@@ -20872,7 +20872,7 @@ void FUN_0045ce90(int param_1)
 #pragma opt_loop_invariants on
 // FUN_0045D990 NONMATCHING
 
-void FUN_0045d990(int param_1,long param_2)
+void FUN_0045d990(int param_1,int param_2)
 
 {
   int iVar1;
